+++ src/Migrations/Version20260616114400.php
@@ -5,6 +5,9 @@ declare(strict_types=1);
 namespace Dbp\Relay\AuthorizationBundle\Migrations;
 
 use Dbp\Relay\AuthorizationBundle\Authorization\AuthorizationService;
+use Dbp\Relay\AuthorizationBundle\Entity\AuthorizationResource;
+use Dbp\Relay\AuthorizationBundle\Entity\AvailableResourceClassAction;
+use Dbp\Relay\AuthorizationBundle\Entity\ResourceActionGrant;
 use Dbp\Relay\AuthorizationBundle\Service\InternalResourceActionGrantService;
 use Doctrine\DBAL\Schema\Schema;
 use Symfony\Component\Uid\UuidV7;
@@ -24,43 +27,52 @@ final class Version20260616114400 extends EntityManagerMigration
         $MANAGE_ITEM_ACTION_UUID = UuidV7::fromString(InternalResourceActionGrantService::MANAGE_ITEM_ACTION_UUID)->toHex();
         $MANAGE_COLLECTION_ACTION_UUID = UuidV7::fromString(InternalResourceActionGrantService::MANAGE_COLLECTION_ACTION_UUID)->toHex();
         $MANAGE_ACTION = AuthorizationService::MANAGE_ACTION;
+        $COLLECTION_RESOURCE_IDENTIFIER = AuthorizationService::COLLECTION_RESOURCE_IDENTIFIER;
+        $ITEM_ACTION_TYPE = AvailableResourceClassAction::ITEM_ACTION_TYPE;
+        $COLLECTION_ACTION_TYPE = AvailableResourceClassAction::COLLECTION_ACTION_TYPE;
 
         $this->addSql('
-             ALTER TABLE authorization_resource_action_grants
+             ALTER TABLE '.ResourceActionGrant::TABLE_NAME.'
              CHANGE action action VARCHAR(40) DEFAULT NULL
         ');
 
         $this->addSql('
-            ALTER TABLE authorization_available_resource_class_actions
-            CHANGE resource_class resource_class VARCHAR(40) DEFAULT NULL
+            ALTER TABLE '.AvailableResourceClassAction::TABLE_NAME.'
+            CHANGE '.AvailableResourceClassAction::RESOURCE_CLASS_COLUMN.' '.AvailableResourceClassAction::RESOURCE_CLASS_COLUMN.' VARCHAR(40) DEFAULT NULL
         ');
 
-        $this->addSql("
-            INSERT INTO authorization_available_resource_class_actions (identifier, resource_class, action, action_type)
+        $this->addSql('
+            INSERT INTO '.AvailableResourceClassAction::TABLE_NAME." (identifier, resource_class, action, action_type)
             VALUES
-                ($MANAGE_ITEM_ACTION_UUID, NULL, '$MANAGE_ACTION', 0),
-                ($MANAGE_COLLECTION_ACTION_UUID, NULL, '$MANAGE_ACTION', 1)
+                ($MANAGE_ITEM_ACTION_UUID, NULL, '$MANAGE_ACTION', $ITEM_ACTION_TYPE),
+                ($MANAGE_COLLECTION_ACTION_UUID, NULL, '$MANAGE_ACTION', $COLLECTION_ACTION_TYPE)
         ");
 
         $this->addSql('
-            ALTER TABLE authorization_resource_action_grants
-            ADD available_resource_class_action_identifier BINARY(16) DEFAULT NULL COMMENT \'(DC2Type:uuid_binary)\'
+            ALTER TABLE '.ResourceActionGrant::TABLE_NAME.'
+            ADD .'.ResourceActionGrant::AVAILABLE_RESOURCE_CLASS_ACTION_IDENTIFIER_COLUMN.' BINARY(16) DEFAULT NULL
         ');
 
         $this->addSql('
-            ALTER TABLE authorization_resource_action_grants
+            ALTER TABLE '.ResourceActionGrant::TABLE_NAME.'
             ADD CONSTRAINT FK_available_resource_class_action_identifier
                 FOREIGN KEY (available_resource_class_action_identifier) REFERENCES authorization_available_resource_class_actions (identifier) ON DELETE CASCADE
         ');
 
-        $this->addSql("
-            UPDATE authorization_resource_action_grants rag
-            JOIN authorization_resources ar
+        $this->addSql('
+            UPDATE '.ResourceActionGrant::TABLE_NAME.' rag
+            JOIN '.AuthorizationResource::TABLE_NAME.' ar
                 ON ar.identifier = rag.authorization_resource_identifier
-            JOIN authorization_available_resource_class_actions arca
-                ON arca.resource_class = ar.resource_class AND arca.action = rag.action
-                OR (arca.resource_class IS NULL AND rag.action = '$MANAGE_ACTION' AND ar.resource_identifier = 'null' AND arca.action_type = 1)
-                OR (arca.resource_class IS NULL AND rag.action = '$MANAGE_ACTION' AND ar.resource_identifier != 'null' AND arca.action_type = 0)
+            JOIN '.AvailableResourceClassAction::TABLE_NAME." arca
+                ON rag.action = arca.action
+                AND (
+                    (ar.resource_identifier = '$COLLECTION_RESOURCE_IDENTIFIER' AND arca.action_type = $COLLECTION_ACTION_TYPE)
+                    OR (ar.resource_identifier != '$COLLECTION_RESOURCE_IDENTIFIER' AND arca.action_type = $ITEM_ACTION_TYPE)
+                )
+                AND (
+                    (arca.resource_class IS NOT NULL AND arca.resource_class = ar.resource_class)
+                     OR (arca.resource_class IS NULL AND rag.action = '$MANAGE_ACTION')
+                )
             SET rag.available_resource_class_action_identifier = arca.identifier
         ");
     }